+++ app/controllers/moneytree/webhooks/stripe_controller.rb
@@ -23,17 +23,22 @@ module Moneytree
       private
 
       def webhook_params
-        @webhook_params ||=
-          ::Stripe::Event.construct_from(
-            JSON.parse(request.body.read, symbolize_names: true)
-          )
+        sig_header = request.env['HTTP_STRIPE_SIGNATURE']
+
+        @webhook_params ||= ::Stripe::Webhook.construct_event(
+          payload,
+          request.env['HTTP_STRIPE_SIGNATURE'],
+          webhook_secret
+        )
+      rescue JSON::ParserError, Stripe::SignatureVerificationError
+        raise ActionDispatch::Http::Parameters::ParseError # Will return status 400
       end
 
       def process_charge!
         return if transaction.completed?
 
         transaction.process_response(
-          Moneytree::TransactionResponse.new(:success, '', { charge_id: stripe_object.id })
+          Moneytree::PspResponse.new(:success, '', { charge_id: stripe_object.id })
         )
 
         if Moneytree.order_status_trigger_method
@@ -51,7 +56,7 @@ module Moneytree
           next if refund.completed?
 
           refund.process_response(
-            Moneytree::TransactionResponse.new(:success, '')
+            Moneytree::PspResponse.new(:success, '')
           )
           if Moneytree.order_status_trigger_method
             transaction.order.send(Moneytree.order_status_trigger_method, transaction)
@@ -60,7 +65,9 @@ module Moneytree
       end
 
       def process_account_updated!
-        payment_gateway = Moneytree.PaymentGateway.find(stripe_object.metadata.moneytree_id.to_i)
+        return if stripe_object.metadata[:payment_gateway_id].blank?
+
+        payment_gateway = Moneytree::PaymentGateway.find(stripe_object.metadata[:payment_gateway_id].to_i)
         confirm_stripe_account(payment_gateway, stripe_object)
       end
 
@@ -70,6 +77,19 @@ module Moneytree
 
       def stripe_object
         @stripe_object ||= webhook_params.data.object
+      end
+
+      def payload
+        @payload ||= request.body.read
+      end
+
+      def webhook_secret
+        case JSON.parse(payload)['type']
+        when 'charge.succeeded', 'charge.refunded'
+          Moneytree.stripe_credentials[:account_webhook_secret]
+        when 'account.updated'
+          Moneytree.stripe_credentials[:connect_webhook_secret]
+        end
       end
     end
   end