+++ lib/moneytree/payment_provider/stripe.rb
@@ -12,31 +12,6 @@ module Moneytree
         super
       end
 
-      def onboarding_url(moneytree_account, current_host)
-        if payment_gateway.psp_credentials&.dig(:account_id)
-          stripe_account = ::Stripe::Account.retrieve(payment_gateway.psp_credentials[:account_id])
-        else
-          stripe_account = ::Stripe::Account.create({
-            type: 'express',
-            capabilities: { card_payments: { requested: true }, transfers: { requested: true } },
-            metadata: { payment_gateway_id: payment_gateway.id, account_id: moneytree_account.id, account_type: moneytree_account.class.name }
-          }.merge(moneytree_account.moneytree_onboarding_data))
-
-          payment_gateway.update! psp_credentials: { account_id: stripe_account.id }
-        end
-
-        stripe_account_link = ::Stripe::AccountLink.create(
-          {
-            account: stripe_account.id,
-            refresh_url: Moneytree::Engine.routes.url_helpers.onboarding_stripe_new_url(host: current_host),
-            return_url: Moneytree::Engine.routes.url_helpers.onboarding_stripe_complete_url(host: current_host),
-            type: 'account_onboarding'
-          }
-        )
-
-        stripe_account_link.url
-      end
-
       def get_access_token(params)
         # FIXME: add error handling
         ::Stripe::OAuth.token(
@@ -51,7 +26,7 @@ module Moneytree
         PERMISSION.to_s
       end
 
-      def charge(amount, details, metadata:, app_fee_amount: 0, description: "Charge for #{account.name}")
+      def charge(amount, details, metadata:, app_fee_amount: 0, description: "Charge for #{transfers.map(&:account_name).join(', ')}}")
         # `source` is obtained with Stripe.js; see https://stripe.com/docs/payments/accept-a-payment-charges#web-create-token
         response = ::Stripe::Charge.create(
           {
@@ -65,7 +40,7 @@ module Moneytree
           stripe_account: payment_gateway.psp_credentials[:stripe_user_id]
         )
         # succeeded, pending, or failed
-        Moneytree::TransactionResponse.new(
+        Moneytree::PspResponse.new(
           { succeeded: :success, pending: :pending, failed: :failed }[response[:status].to_sym],
           response[:failure_message],
           {
@@ -75,7 +50,7 @@ module Moneytree
           }
         )
       rescue ::Stripe::StripeError => e
-        Moneytree::TransactionResponse.new(:failed, e.message)
+        Moneytree::PspResponse.new(:failed, e.message)
       end
 
       def refund(amount, details, metadata:)
@@ -90,13 +65,13 @@ module Moneytree
         )
 
         # succeeded, pending, or failed
-        Moneytree::TransactionResponse.new(
+        Moneytree::PspResponse.new(
           { succeeded: :success, pending: :pending, failed: :failed }[response[:status].to_sym],
           response[:failure_message],
           { refund_id: response[:id] }
         )
       rescue ::Stripe::StripeError => e
-        Moneytree::TransactionResponse.new(:failed, e.message)
+        Moneytree::PspResponse.new(:failed, e.message)
       end
 
       def card_for(transaction)