+++ test/dummy/app/models/merchant_order.rb
@@ -1,0 +1,6 @@
+class MerchantOrder < ApplicationRecord
+  include Moneytree::AccountOrder
+
+  belongs_to :order
+  belongs_to :merchant
+end